% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_empty_result}
\alias{create_empty_result}
\title{Create empty result data frame}
\usage{
create_empty_result(end_year, type)
}
\arguments{
\item{end_year}{School year end}

\item{type}{Entity type ("State", "District", or "School")}
}
\value{
Empty data frame with expected columns
}
\description{
Create empty result data frame
}
\keyword{internal}
