% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_mde_enrollment}
\alias{download_mde_enrollment}
\title{Download enrollment data from MDE portal}
\usage{
download_mde_enrollment(end_year, level = "school")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Entity level: "school" or "district"}
}
\value{
Data frame with enrollment data
}
\description{
Downloads data from the Mississippi Department of Education's
newreports.mdek12.org data portal.
}
\keyword{internal}
