% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_mde_alternative}
\alias{fetch_mde_alternative}
\title{Alternative download method for MDE data}
\usage{
fetch_mde_alternative(end_year, entity_type)
}
\arguments{
\item{end_year}{School year end}

\item{entity_type}{"School" or "District"}
}
\value{
Data frame with enrollment data
}
\description{
Uses the Data Download page which provides Excel/CSV exports
}
\keyword{internal}
