% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_mde_api_data}
\alias{fetch_mde_api_data}
\title{Fetch data from MDE API}
\usage{
fetch_mde_api_data(end_year, entity_type)
}
\arguments{
\item{end_year}{School year end}

\item{entity_type}{"School" or "District"}
}
\value{
Data frame with enrollment data
}
\description{
Fetch data from MDE API
}
\keyword{internal}
