% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_mde_static_files}
\alias{fetch_mde_static_files}
\title{Fetch MDE static data files}
\usage{
fetch_mde_static_files(end_year, entity_type)
}
\arguments{
\item{end_year}{School year end}

\item{entity_type}{"School" or "District"}
}
\value{
Data frame with enrollment data
}
\description{
Downloads pre-generated data files from MDE servers
}
\keyword{internal}
