% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get list of available years}
\usage{
get_available_years()
}
\value{
Integer vector of available end years
}
\description{
Returns the range of school years for which Mississippi enrollment
data is available.
}
\examples{
get_available_years()
}
