% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_mde_column_map}
\alias{get_mde_column_map}
\title{Get column mapping for MDE data}
\usage{
get_mde_column_map()
}
\value{
Named list of column mappings
}
\description{
Returns a mapping of MDE column names to standardized names.
MDE column names vary slightly across years.
}
\keyword{internal}
